% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud2NA.R
\name{cloud2NA}
\alias{cloud2NA}
\title{Cloud To NA}
\usage{
cloud2NA(x, y)
}
\arguments{
\item{x}{rastwe layer to be masked, ie the original}

\item{y}{raster layer used as the mask, all pixels not 0 will be NA, potentially use reclass to}
}
\value{
a multiband tiff image where non 0 pixels have been removed (NA'D)
}
\description{
Function to mask out all areas classed as cloud/cloud shadow/water/adjacent cloud by the CFmask layer provided by CDR
Landsat imagery.
}

