% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_removal.R
\name{cloud_removal}
\alias{cloud_removal}
\title{Remove Clouds in Landsat Imagery}
\usage{
cloud_removal(path, outfolder)
}
\arguments{
\item{path}{path to a multiband tiff image produced by auto_preprocess_landsat in the team_lucc package}

\item{outfolder}{folder where the output image will be saved, file name will be the same as input}
}
\value{
a multiband tiff image where non "clear" pixels have been removed (NA'D)
}
\description{
Function to mask out all areas classed as cloud/cloud shadow/water/adjacent cloud by the CFmask layer provided by CDR
Landsat imagery. This is the final stage of landsat preprocessing workflow. (espa_extract, auto_preprocess_landsat).
SR and mask files should be in the same folder.
}

