% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{lassie_get}
\alias{lassie_get}
\title{Return the value of 'lassie' object}
\usage{
lassie_get(x, what_x)
}
\arguments{
\item{x}{\code{\link[zebu]{lassie}} S3 object.}

\item{what_x}{vector specifying values to be returned:
\itemize{
\item 'local': local association measure values (default).
\item 'obs': observed probabilities.
\item 'exp': expected probabilities.
\item 'local_p': p-value of local association (after running \code{\link[zebu]{permtest}} or \code{\link[zebu]{chisqtest}}).
}}
}
\value{
Corresponding array contained in \code{\link[zebu]{lassie}} object.
}
\description{
Subroutine for \code{\link[zebu]{lassie}} methods. Tries to retrieve a value from a \code{\link[zebu]{lassie}} object
and gives an error if value does not exist.
}
\examples{
las <- lassie(trees)
las_array <- lassie_get(las, 'local')

}
