% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Landsat_LST.R
\name{LST}
\alias{LST}
\title{Landsat-based Land Surface Temperature}
\usage{
LST(red, nir, bt, ex, fmask, band, output_prefix)
}
\arguments{
\item{red,}{red waveband reflectance, Landsat4/5/7= band 3, Landsat 8=Band 4}

\item{nir}{near infared waveband reflectance, Landsat 4/5/7=band 4, Landsat 8=Band 5}

\item{bt}{Brightness temperature band. Landsat 4/5/7=Band 6, Landsat 8= bands 10 or 11}

\item{ex}{extent of processing}

\item{band}{thermal band used, Landsat 4/5/6= "TM", Landsat 8 "TIRS10" or "TIRS11" depending on band}

\item{output_prefix}{prefix for output file eg "Manchester_2003_july10" LST/ndvi.tif will be added for file}
}
\value{
a stack containing LST and NDVI

writes NDVI and LST rasters to the current wd
}
\description{
Landsat-based Land Surface Temperature
}

