% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PalsarBackscatter.R
\name{PalsarBackscatter}
\alias{PalsarBackscatter}
\title{Palsar Mosaic to backscatter}
\usage{
PalsarBackscatter(palsar)
}
\arguments{
\item{palsar,}{a raster object (layer,stack,or brick) of ALOS PALSAR mosaic}
}
\value{
A corresponding raster object with the DN converted to backscatter
}
\description{
Converts ALOS PALSAR global mosaic digital numbers (DN) to backscatter
}

