% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{calculate the standard error of the mean}
\usage{
se(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
calculatess the standard error of the mean for a vector.
}
\examples{
a <- rnorm(1000, 20, 5)
mean(a)
se(a)

}
\author{
Tom Higginbottom
}

