% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_df.R
\name{split_df}
\alias{split_df}
\title{Split Data Frame}
\usage{
split_df(dataframe, seed = NULL)
}
\arguments{
\item{dataframe}{The dataframe to be split}

\item{seed}{Option to set seet (random number generator) for reproducability, defaults to null}
}
\value{
A list of training and validation dataframes, extract usiing training  <- input$trainset or input$testset
}
\description{
Function to randomly split a data frame into two sections, potentially for training and validation
As R will only allow a single item to be returned by a functio use
}

