\name{zebu-package}
\alias{zebu-package}
\alias{zebu}
\docType{package}
\title{
\packageTitle{zebu}
}
\description{
\packageDescription{zebu}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{zebu}
\packageIndices{zebu}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{zebu}

Maintainer: \packageMaintainer{zebu}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
