% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zebu.R
\docType{package}
\name{zebu}
\alias{zebu}
\title{zebu: Local Association Measures}
\description{
The zebu package implements the estimation of local (and global) association measures: Ducher's Z, pointwise mutual information and normalized pointwise mutual information. The significance of local (and global) association is accessed using p-values estimated by permutations.
}
\section{Functions}{


\code{\link[zebu]{lassie}} estimates local (and global) association measures: Ducher's Z, pointwise mutual information and normalized pointwise mutual information.

\code{\link[zebu]{permtest}} accesses the significance of local (and global) association values using  p-values estimated by permutations.
}

